More cleanup
This commit has a few bugs with the movement of the characters due to
timing and also removed some of the features of the early game like
morphing multiple lives, this so it will be simpler to get finished and
working.

+++ src/Animals.h
@@ -10,7 +10,7 @@
 class AnimalState
 {
 public:
-	int xPos, yPos;
+	float xPos, yPos;
 	int Frame, State;
 	int Height, Width;
 	int Radius;
